pyproject.toml
[project]
name = "tomato-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.116.1",
    "sqlalchemy>=2.0.43",
    "pydantic-settings>=2.4.0",
    "uvicorn[standard]>=0.30.1",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "python-telegram-bot>=21.0.1",
    "schedule>=1.2.2",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.2",
]
